+++ src/sinks/aws_sqs/mod.rs
@@ -1,4 +1,5 @@
 mod config;
+mod request_builder;
 mod retry;
 mod service;
 mod sink;
